+++ packages/create/templates/dbos-drizzle/drizzle/0000_dbos.sql
@@ -1,0 +1,10 @@
+CREATE SCHEMA IF NOT EXISTS dbos;
+
+CREATE TABLE IF NOT EXISTS dbos.transaction_completion (
+    workflow_id TEXT NOT NULL,
+    function_num INT NOT NULL,
+    output TEXT,
+    error TEXT,
+    created_at BIGINT NOT NULL DEFAULT (EXTRACT(EPOCH FROM now())*1000)::bigint,
+    PRIMARY KEY (workflow_id, function_num)
+);
