+++ packages/create/templates/dbos-drizzle/package.json
@@ -13,7 +13,7 @@
     "@dbos-inc/eslint-plugin": "^3.3.4",
     "@types/jest": "^29.5.12",
     "@types/supertest": "^2.0.16",
-    "drizzle-kit": "^0.30.1",
+    "drizzle-kit": "^0.31.4",
     "eslint": "^8.57.0",
     "jest": "^29.7.0",
     "nodemon": "^3.1.0",
@@ -23,6 +23,8 @@
   },
   "dependencies": {
     "@dbos-inc/dbos-sdk": "file:../../../..",
-    "drizzle-orm": "^0.38.3"
+    "@dbos-inc/drizzle-datasource": "file:../../../drizzle-datasource",
+    "drizzle-orm": "^0.44.3",
+    "express": "^4.21.2"
   }
 }
