+++ packages/create/templates/dbos-drizzle/src/main.test.ts
@@ -1,11 +1,10 @@
 import { DBOS } from '@dbos-inc/dbos-sdk';
-import { Hello } from './operations';
+import { app, Hello } from './main';
 import request from 'supertest';
 
 describe('operations-test', () => {
   beforeAll(async () => {
     await DBOS.launch();
-    await DBOS.launchAppHTTPServer();
   });
 
   afterAll(async () => {
@@ -28,7 +27,7 @@ describe('operations-test', () => {
    * Test the HTTP endpoint.
    */
   test('test-endpoint', async () => {
-    const res = await request(DBOS.getHTTPHandlersCallback()).get('/greeting/dbos');
+    const res = await request(app).get('/greeting/dbos');
     expect(res.statusCode).toBe(200);
     expect(res.text).toMatch('Hello, dbos! We have made');
   });
