+++ packages/create/templates/dbos-prisma/src/main.test.ts
@@ -5,7 +5,6 @@ import request from 'supertest';
 describe('operations-test', () => {
   beforeAll(async () => {
     await DBOS.launch();
-    await DBOS.launchAppHTTPServer();
   });
 
   afterAll(async () => {
